Move plain object serialization to separate file
